backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.10.3
pydantic[email]==1.10.7
python-dotenv==1.0.0
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'
pillow==9.5.0
numpy==1.24.2
pandas==2.0.0
scikit-learn==1.2.2
tensorflow==2.12.0
redis==4.5.5
celery==5.2.7
gunicorn==20.1.0
python-memcached==1.59
pytest==7.3.1
httpx==0.23.3
requests==2.31.0
